Access tokens with zero expiry are not expired
Lets us create a Client with an App API Key for which we don't know the
expiry. This is necessary because when we create or rotate an App API
Key, we need to use the newly created key to get its own creation date
(thus expiry). CreateAppAPIKey and RotateAppAPIKey don't return this.
Another function, GetAppAPIKeyMetadata, must be called.